refactoring, added tests, fixed minor bug
