+++ nautilus/bounds/__init__.py
@@ -1,8 +1,10 @@
 """Modules implementing various multi-dimensional bounds."""
 
 from .basic import UnitCube, Ellipsoid, UnitCubeEllipsoidMixture
-from .neural import NeuralBound, NautilusBound
+from .nautilus import NautilusBound
+from .neural import NeuralBound
+from .periodic import PhaseShift
 from .union import Union
 
-__all__ = ['UnitCube', 'Ellipsoid', 'UnitCubeEllipsoidMixture', 'Union',
-           'NeuralBound', 'NautilusBound']
+__all__ = ['Ellipsoid', 'NautilusBound', 'NeuralBound',
+           'PhaseShift', 'Union', 'UnitCube', 'UnitCubeEllipsoidMixture']
